+++ src/Modules/ElementUtils/ArrayUtils.js
@@ -1,0 +1,4 @@
+function isLastArrayItem(arrayElement, currentArrayItemIndex) {
+  const numberOfArrayItems = arrayElement.content.length;
+  return (numberOfArrayItems - 1) === currentArrayItemIndex;
+}
