+++ src/Modules/ElementUtils/DefaultValueUtils.js
@@ -1,0 +1,11 @@
+function hasDefaults(element) {
+  if (!element) {
+    return false;
+  }
+
+  if (!element.attributes) {
+    return false;
+  }
+
+  return !!element.attributes.default;
+}
