+++ src/Modules/ElementUtils/DescriptionUtils.js
@@ -1,0 +1,9 @@
+function hasDescription(element) {
+  if (element.meta) {
+    if (element.meta.description) {
+      return true;
+    }
+  }
+
+  return false;
+}
