+++ src/Modules/ElementUtils/QueryUtils.js
@@ -1,0 +1,19 @@
+function findElement(elementId, dataStructures) {
+  const query = {
+    'meta.id': elementId,
+  };
+
+  const results = sift(query, dataStructures);
+
+  if (isEmpty(results)) {
+    return null;
+  }
+
+  return results[0];
+}
+
+function findParent(elementId, dataStructures) {
+  const element = findElement(elementId, dataStructures);
+
+  return findElement(element.meta.ref, dataStructures);
+}
