+++ src/Modules/ElementUtils/SamplesUtils.js
@@ -1,0 +1,11 @@
+function hasSamples(element) {
+  if (!element) {
+    return false;
+  }
+
+  if (!element.attributes) {
+    return false;
+  }
+
+  return !isEmpty(element.attributes.samples);
+}
